+++ core/src/main/java/org/svip/sbomfactory/generators/generators/cyclonedx/CycloneDXStore.java
@@ -47,7 +47,7 @@ public class CycloneDXStore extends BOMStore {
      * @param bomVersion The version of the SBOM - 1 if the first one generated, n if the nth one generated.
      * @param headComponent The head component of the SBOM. This is the component that stores the SBOM name, licenses, etc
      */
-    public CycloneDXStore(String serialNumber, int bomVersion, ParserComponent headComponent) {
+    public CycloneDXStore(String serialNumber, Integer bomVersion, ParserComponent headComponent) {
         super(GeneratorSchema.CycloneDX, "1.4", serialNumber, bomVersion, headComponent);
 
         this.components = new ArrayList<>();